Implemented GetAllAsync

+++ src/Infrastructure/Presistance/Repositories/IShortcutRepository.cs
@@ -10,7 +10,7 @@ namespace Presistance.Repositories
         public Task<Shortcut> FindByAliasAsync(string alias);
         public Task<long> GetCountAsync();
         public Task<List<Shortcut>> GetAllAsync();
-        public Task<List<Shortcut>> GetAllAsync(long take, long skip);
+        public Task<List<Shortcut>> GetAllAsync(int take, int skip);
         public Task InsertAsync(Shortcut shortcut);
         public Task<bool> DeleteAsync(long id);
         public Task UpdateAsync(Shortcut shortcut);
